replaced TriangleMesh code with new TriangleMesh
